Make indented blocks work
We were misdetecting the end of the docblock because our `end` regex
could match lines like:

    <space><space>"

(ie. a 0 or 1-width space, followed by a non-quote)

Fix that by looking for non-space, non-tab, non-quote instead.

Also, use "\s" for whitespace, so that we deal with tabs properly as
well as spaces.