Italicize all the things
Comments are usually italics, so these should be too.

Note that I can't actually ship this as it depends on
`functions#italicize_group`, which is only defined in my dot files, but
I will eventually extract that too into a separate plug-in so that I can
re-use it here too.